DVT-#115 이미지 마커 오버레이를 만드는 로직을 함수로 분리
